Update README.md
I am a student at FGCU university, as of the time of this writing I am it is 3/12/2020 and I am in my "Intro To Computer Science" class. My professor has asked me to make a GitHub account and I am testing the website by making this repository. I have been practicing python for about three months now, but I still have a rudimentary understanding of it.